+++ Johnson/docs/release-notes.md
@@ -4,6 +4,7 @@
 
 ## Upcoming version
 
+* Layout and Strategize now put the cards in the discard pile instead of the draw pile.
 * Fixed Strengthen not updating text card attacks visually.
 
 ## 1.0.2